% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{angle2coord}
\alias{angle2coord}
\title{Angular coordinates}
\usage{
angle2coord(angles, cen = c(0, 0), r = 1)
}
\arguments{
\item{angles}{angles in radians}

\item{cen}{numeric, x and y coordinates to locate the center of the circle}

\item{r}{numeric, radius of the circle}
}
\value{
Angular coordinate matrix based on a set of angles
}
\description{
Given a circle, it calculates where in the circumference will a line
at a certain angle cut through the circumference.
}
\keyword{internal}
