% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{axis_wheel}
\alias{axis_wheel}
\title{Axis wheel}
\usage{
axis_wheel(res, cen = c(0, 0), r = 1, ...)
}
\arguments{
\item{cen}{numeric, x and y coordinates to locate the center of the circle}

\item{r}{numeric, radius of the circle}

\item{...}{other parameters to pass to segments}

\item{n}{resolution of the circle}
}
\value{
plots the axis of around a circle
}
\description{
Draws axis ticks (not circle)
}
\keyword{internal}
