% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{calc.Hinv}
\alias{calc.Hinv}
\title{Calculates \eqn{H^-1} matrix}
\usage{
calc.Hinv(y, X, Z, K = NULL, bounds = c(1e-09, 1e+09), method = "REML")
}
\arguments{
\item{y}{Response vector or matrix. Each column is taken as a different response.}

\item{X}{Fixed effects design matrix}

\item{Z}{Random effects design matrix}

\item{K}{Optional (defaults to NULL).
Variance-covariance square matrix (Genetic distance matrix).}

\item{bounds}{Bounds of the Ridge Regression parameter}

\item{method}{"REML" or "ML". If "ML", \eqn{H^-1} is calculated using maximum likelihood
equations. If REML, Restricted Maximum Likelihood.}
}
\value{
Returns a list \eqn{H^-1},
where each element corresponds to the \eqn{H^-1} of each column of y. An
attribute "lambda" is included in each \eqn{H^-1} matrix that contains the optimization
paramater of ridge regression.
}
\description{
The core of a Ridge-Regression solution for a mixed model
is the calculation of a Hat matrix, or projection matrix \eqn{H}, which contains the
variance components of the random part of the model. In the EMMAX/P3D approach,
the inverse of such matrix, \eqn{H^-1}, is calculated only once and recycled for the analysis
of each marker, which obtains an approximate result but speeds up computational time
considerably. For more information consult the articles that simulatenously presented this method:
\href{https://www.nature.com/articles/ng.546}{P3D: Zhang et al. 2010}
\href{https://www.nature.com/articles/ng.548}{EMMAX: Kang et al. 2010}
}
\keyword{internal}
