% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{calc.K}
\alias{calc.K}
\title{Calculation of realized distance matrix (K)}
\usage{
calc.K(matrix, haplotypes = F, ploidy = NULL)
}
\arguments{
\item{matrix}{Numeric matrix with individuals on rows and markers on columns.}

\item{haplotypes}{logical, whether haplotypes are present in the matrix. If T,
a matrix is expected to have p columns per individual, where p is ploidy.}

\item{ploidy}{integer, the ploidy number p of all individuals. Only used if
haplotypes = T.}
}
\value{
A numeric matrix nxn where n is the number of rows.
}
\description{
Using dosage or haplotype scores, a distance matrix is calculated
such that the average distance of an individual with itself is 1, and
the average with an unrelated individual is 0. Based on the "Realized
Relationship" model found in \href{https://dl.sciencesocieties.org/publications/tpg/abstracts/9/2/plantgenome2015.08.0073}{Rosyara et al. 2016}
}
\examples{
#Create 10 tetraploid individuals with 200 markers each
inds <- lapply(1:10,function(i) round(runif(200,0,4)) )

#Put them in a matrix or data.frame with individuals in rows
geno <- do.call(rbind,inds)

#K matrix of 10 individuals (each row is one individual)
K<-calc.K(geno)

#K matrix of 2 individuals (every 5 rows is one individual)
K <- calc.K(geno, haplotye = T, ploidy = 5)

}
