% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{calc.Q}
\alias{calc.Q}
\title{Calculate a Q design matrix based on a vector}
\usage{
calc.Q(pop, names = NULL)
}
\arguments{
\item{pop}{Vector where each element is a population identifier}

\item{names}{Optional. A vector of names for the \eqn{Q} matrix}
}
\value{
A normalized matrix with as many columns as population groups
[ncol=unique(pop)] identifying each individual belonging to
one population.
}
\description{
Using a vector of values as input, it creates a normalized cofactor
design matrix (\eqn{Q}) that identifies each individual as belonging to a group.
}
\keyword{internal}
