% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{circle}
\alias{circle}
\title{Circle drawing}
\usage{
circle(cen = c(0, 0), r = 1, res = 1200, ...)
}
\arguments{
\item{cen}{numeric, x and y coordinates to locate the center of the circle}

\item{r}{numeric, radius of the circle}

\item{res}{numeric, resolution of the cirlce, over 60 recommended}

\item{...}{other parameters to pass to segment}
}
\value{
plots a circle
}
\description{
Circle drawing
}
\keyword{internal}
