% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{colour_wheel}
\alias{colour_wheel}
\title{Colour wheel}
\usage{
colour_wheel(res, cen = c(0, 0), r = 1, l = NULL, ...)
}
\arguments{
\item{res}{Number of sections of different colour to draw in the circle}

\item{cen}{numeric, x and y coordinates to locate the center of the circle}

\item{r}{numeric, radius of the circle}

\item{l}{numeric, values of luminance for the colours}

\item{...}{}
}
\value{
plots a colour wheel based on triangles
}
\description{
Draws a colour wheel with an indicated resolution
}
\keyword{internal}
