% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{comp.vec}
\alias{comp.vec}
\title{Vector comparison function}
\usage{
comp.vec(vec1, vec2)
}
\arguments{
\item{vec1}{vector of values}

\item{vec2}{vector of values}
}
\value{
Matrix of logical values indicating whether the values of vec1
and vec2 are equal. Columns correspond to elements
of vec1 and rows to elements of vec2.
}
\description{
Element per element comparison between two vectors.
}
\keyword{internal}
