% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{dosage.X}
\alias{dosage.X}
\title{X matrix calculator}
\usage{
dosage.X(genotypes, haplotype = F, ploidy = NULL, normalize = F)
}
\arguments{
\item{genotypes}{vector of dosages per individual, or p haplotypes per individual
(where p is ploidy).}

\item{haplotype}{logical, whether the vector contains haplotype or genotype data}

\item{ploidy}{integer indicating ploidy.}

\item{normalize}{logical, should the X matrix be normalized?}
}
\value{
if haplotypes = F, a matrix is returned whith a single column
containing the same values that were provided. If haplotypes = T, a design matrix
is created with ncol = number of haplotypes and nrow = number of individuals.
}
\description{
Using a vector of dosages per individual, or p haplotypes per
individual (where p is ploidy), it creates a design matrix X, with or without
normalization.
}
\examples{

dos <- round(runif(100,0,4))
dosage.X(dos)

hap <- round(sample(400,1:20))
dosage.X(haplotype,haplotype = T, ploidy = 4)
}
