% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{draw_chrom_axis}
\alias{draw_chrom_axis}
\title{Chromosome axis drawer}
\usage{
draw_chrom_axis(ch_edges, small = F)
}
\arguments{
\item{ch_edges}{data.frame with first column having the axis start of each
chromosome, and the second column the axis end. Row names must have the
names of each chromosome. This data.frame can be obtained using \code{map_axis}}

\item{small}{logical, should small axis be also drawn?}
}
\value{

}
\description{

}
\keyword{internal}
