% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fun.R
\name{HapCurate}
\alias{HapCurate}
\title{Prepare haplotypes for QTL mapping}
\usage{
HapCurate(
  haplo,
  hb_list,
  map = NULL,
  ploidy,
  na.rate = 1,
  use.SNPs = F,
  snpdose
)
}
\arguments{
\item{haplo}{It can be one of the following options:
\itemize{
  \item A matrix of multi-allelic genotypes ('haplotype names'), with
  markers in rows and individual homologues in columns. Marker names
  are provided as rownames and individual homologues names as column names.
  Genotypes can be both numbers and characters.
  \item A list where each element
  is a numeric matrix of haplotype dosages for a certain locus and each element name
  is a locus name. Locus names must be unique.
  In each matrix, haplotypes are in rows and individuals in columns.
  Row names contain haplotype names.
  Column names contain individual names.
  \item The output of PolyHaplotyper.
}}

\item{hb_list}{A list where each element is a vector of SNP names contained
in a block.}

\item{map}{A data frame with three columns named "marker", "chromosome"
and "position".}

\item{ploidy}{Numeric indicating the ploidy level.}

\item{na.rate}{Numeric, from 0 to 1, indicating the rate of missingness
allowed per marker (or block). Using 1 no filtration is applied.}

\item{use.SNPs}{Logical value. If TRUE, SNP markers of discarded blocks will be
included in the final haplotype table. Default is FALSE.}

\item{snpdose}{A numeric matrix of SNP dosages, with markers in rows and
individuals in columns. Row names are marker names and colmn names are
individual names.}
}
\value{
A list with two elements
\itemize{
  \item *$genotypes* Curated matrix of multi-allelic genotypes
  ('haplotype names'), with markers in rows and individual homologues in
  columns. Marker names are provided as rownames and individual homologues
  names as column names. Genotypes can be both numbers and characters.
  \item *$map* A map for the haplotypes, where the position of each marker
  is the average position of the SNP markers contained in a block.
}
}
\description{
Prepare haplotypes for QTL mapping
}
