% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fun.R
\name{HapnameToHapdose}
\alias{HapnameToHapdose}
\title{Convert haplotype names into haplotype dosages}
\usage{
HapnameToHapdose(m)
}
\arguments{
\item{x}{a matrix of haplotype names, with markers in rows and individual
homologues in columns. Column names for homologues of the same individual
are distinguished by adding the suffices "_1","_2", etc., depending on
the ploidy.}
}
\value{
a list of matrices. Each matrix contains haplotype dosages for one
marker (or block), with haplotypes in rows and individuals in columns.
}
\description{
A genotype of an individual can be represented either by a
numerical vector of dosages of each haplotype or by a numerical vector of
haplotype names (numbers) for each homologue. In the former case, vector
length will correspond to the number of haplotypes, while in the latter
case it will be the number of homologues, that is the individual ploidy.
This function converts a matrix of haplotypes names into a list of
haplotype dosages.
}
