% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{hue_wheel}
\alias{hue_wheel}
\title{Hue wheel}
\usage{
hue_wheel(l = NULL)
}
\arguments{
\item{l}{luminance parameter, which can be modified to obtain lighter/darker
colours. Values between 20 and 100 are recommended (above and below
not all colours exist). Defaults to 60.}
}
\value{
plots a hue wheel of a specific luminance
}
\description{
Draws the hue wheel of colorspace in order to know what number corresponds to each
colour
}
\examples{

hue_wheel() #luminance = 60 by default
hue_wheel(90) #luminance = 90
}
