% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{imputeNA}
\alias{imputeNA}
\title{NA replacement by mean score per marker}
\usage{
imputeNA(m, miss = NULL)
}
\arguments{
\item{m}{a numeric matrix, with markers in rows and
individuals in columns.}

\item{miss}{a number used for missing values instead of NA.
The function expects missing values are indicated by NA. Alternatively,
since the inpute matrix is a numeric matrix, you could indicate
missing values by a number.}
}
\value{
a matrix with no missing values.
}
\description{
For some operations (e.g. calculating correlation) the
presence of missing values is not allowed. If the level of missingness
per marker is not too high (below 25%), missing values can be imputed
using the mean score per marker, so that allele frequency is not affected.
}
\keyword{internal}
