% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{inputCheck_K}
\alias{inputCheck_K}
\title{Check input requirements for the argument K of map.QTL}
\usage{
inputCheck_K(K)
}
\arguments{
\item{K}{the input of the argument K of map.QTL}
}
\value{
Nothing. Stop with error if requirements are not met.
}
\description{
Check that the argument K is either i) NULL or ii) TRUE or
iii) a numeric square matrix.
}
\keyword{internal}
