% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{inputCheck_Q}
\alias{inputCheck_Q}
\title{Check input requirements for the argument Q of map.QTL}
\usage{
inputCheck_Q(Q)
}
\arguments{
\item{K}{the input of the argument Q of map.QTL}
}
\value{
Nothing. Stop with error if requirements are not met.
}
\description{
Check that the argument Q is either i) NULL or ii) TRUE or
iii) a vector.
}
\keyword{internal}
