% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{inputCheck_snp}
\alias{inputCheck_snp}
\title{Check format of a numeric input matrix}
\usage{
inputCheck_snp(x, integer = TRUE, ploidy = NULL)
}
\arguments{
\item{x}{a matrix or a data frame.}

\item{integer}{Logical value indicating whether to check that the input
matrix contains integers only (e.g. allelic dosages). Default if
integer = TRUE. If FALSE, non-integers are allowed and the ploidy check is
skipped.}

\item{ploidy}{A number specifying the ploidy level. If a ploidy level is
provided, the function will check whether all the data fit with that ploidy.}
}
\value{
a matrix
}
\description{
Used in map.QTL and thr.LiJi
}
\keyword{internal}
