% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{LD_decay}
\alias{LD_decay}
\title{LD decay calculation}
\usage{
LD_decay(
  dos,
  map,
  win_size = 0.1,
  max_dist = NULL,
  per_chr = F,
  percentile = c(0.5, 0.8, 0.9, 0.95)
)
}
\arguments{
\item{dos}{integer matrix with markers on rows and individuals in columns}

\item{map}{data.frame genetic map with columns "chromosome" "marker" and "position"}

\item{win_size}{numeric indicating the window size to calculate LD estimates.
Must be in the same unit as the "position" column of map.}

\item{max_dist}{numeric, maximum distance to consider between markers. Usually, beyond
50 cM the estimates are not informative.}

\item{per_chr}{logical, whether the estimations should be performed per chromosome}

\item{percentile}{numeric vector, which percentiles to calculate?  By default
0.5, 0.8, 0.9 and 0.95. Any value between 0 and 1 is allowed.}
}
\value{
If per_chr = F, an LD object (list with LD estimates dataframe and
background LD for each percentile); if per_chr = T a list of LD objects, one
per chromosome.
}
\description{
Calculation of linkage disequilibrium per windows across
the genome can help paint a picture of the distribution of recombinations
in a population. This function, provided a set of dosages and a genetic
map, is able to obtain LD estimates for any percentile, by default
at 50%, 80% 90% and 95%. It also calculates the background LD (between unlinked
markers) and can be performed over the whole genome, or per chromosome.
}
\examples{

dosage <- sapply(1:50,function() sample(100,0:4,replace = T))
map <- data.frame(marker = paste0("marker",1:100),
                  chromosome = c(rep(1,30),rep(2,40),rep(3,30)),
                  position = c(1:30,1:40,1:30))
LD <- LD_decay(dosage,map,win_size = 0.5)
}
