% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{lighten}
\alias{lighten}
\title{Colour lightening}
\usage{
lighten(col, factor = 0.55)
}
\arguments{
\item{col}{One or more colours.}

\item{lighten}{Lightening factor as a value between 0 and 1 (default 0.55).
If negatives are used, it darkens the colour.}
}
\value{
Lightened (or darkened) colour(s)
}
\description{
Returns a lighter (if lighten parameter is positive) version of the colour(s) provided.
}
\examples{
\dontrun{
lighten("black")
lighten("#000000")
lighten("red",-0.55) #this returns a darker colour
}

}
\keyword{internal}
