% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{lm_compare}
\alias{lm_compare}
\title{Linear model comparison}
\usage{
lm_compare(X, N, y)
}
\arguments{
\item{X}{matrix of fixed effects in the full model}

\item{N}{matrix of fixed effects in the naive model}

\item{y}{matrix or vector of phenotypes}
}
\value{
list with beta (fixed effects), Ftest (F statistic)
pval (p-value of the difference between models) and se (standard error
of fixed effect estimates)
}
\description{
This function allows to perform a comparison test
between two linear models. This is useful to test groups of fixed
effects. In our program, this is applied to test the "genetic effects"
without testing the cofactors. For instance, if we have a model such as:
y = c1 + c2 + g1 + g2 + e, where c1 and c2 are cofactors, and g1 and
g2 are genetic effects, we would like to know what is the explanatory
value of only g1 and g2. To do so, we compare the model above with
the following model: y = c1 + c2 + e. If g1 and g2 add explanatory value,
then the first model should have higher explanatory value. In this
function, the first model, or full model, is represented by the
matrix X, while the second model, or naive model, is represented by
the matrix N.
}
\keyword{internal}
