% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fun.R
\name{map2blocks}
\alias{map2blocks}
\title{Group SNPs into haploblocks based on a genetic map}
\usage{
map2blocks(map, winsize, sldpace = winsize)
}
\arguments{
\item{map}{a data frame with three columns named "marker", "chromosome"
and "position".}

\item{winsize}{numeric specifing the maximum block size (in cM).}

\item{sldpace}{numeric indicating the minimum distance between consecutive.
blocks}
}
\value{
a list where each element contains the names of markers grouped
in a block
}
\description{
Group SNPs into haploblocks based on a genetic map
}
