% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{map_axis}
\alias{map_axis}
\title{Axis calculator}
\usage{
map_axis(map, space = 0, maxes = NULL)
}
\arguments{
\item{map}{map dataframe (with position, chromosome and marker)}

\item{space}{space to leave between chromosomes (in cM)}

\item{maxes}{the maximum length of each chromosome}
}
\value{
a matrix with marker positions in a map and on an axis
}
\description{
Using a map dataframe, it calculates the
position of each marker onto a plot, including
spaces between chromosomes, if desired.
}
\keyword{internal}
