% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{max_map}
\alias{max_map}
\title{Chromosome maximum position}
\usage{
max_map(map, chr = NULL)
}
\arguments{
\item{map}{data.frame containing a genetic map (at least a chromosome and
a position column)}

\item{chr}{optional, vector with chromosome(s) name(s) to return the
maximums of.}
}
\description{
Given a genetic map (data.frame with columns chromosome
and position), returns the maximum position per chromosome.
}
\keyword{internal}
