% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{mm.solve}
\alias{mm.solve}
\title{Mixed Model Solver}
\usage{
mm.solve(y, X, Z, K, Hinv = NULL, random = FALSE, no.test = 0)
}
\arguments{
\item{y}{Numeric vector of response variable}

\item{X}{Fixed effect design matrix. Must include an intercept. Generally this
matrix corresponds to 1) cofactors and 2) marker(s) genotype(s).}

\item{Z}{Random effects design matrix. Generally this matrix relates observations with
each genotype/individual.}

\item{K}{Variance-covariance structure. If not specified, \code{\link{calc.K}} is used}

\item{Hinv}{\eqn{H^-1} matrix. If not provided \code{\link{calc.K}} will be used. Defaults
to NULL.}

\item{random}{Logical value indicating whether random effects should be returned (slows
down function significantly).}

\item{no.test}{Number of parameters in the X matrix that will not be taken into
account in the pvalue calculation. In practice, this means that the first \code{no.test}
columns of the X matrix will not be used to calculate p-value. For instance, if 3
cofactors are added, the X matrix should have 5 columns: one intercept, three cofactors
and one genotype column. If we wish to investigate the effect of the genotype
only, no.test should take value 4. If we would set \code{no.test} to 1, the p-value
would reflect whether the cofactors have any influence on the response \eqn{y}.}
}
\value{
a data.frame where each column corresponds to a column of \eqn{y}. There are six
rows on the df: beta (estimates of parameters), Fstat (approximation of
F value), residual (vector of residuals), pval (p-value), se (standard error of the model),
wald (Wald test values), and real.df (realized degrees of freedom, df approximation).
If \code{random} is set to \code{True} an extra row "random" is added, containing the
random terms -at the cost of computational efficiency.
}
\description{
Mixed model solver. This function is partly based on the
\code{mixed.solve} function from the rrBLUP package \href{https://doi.org/10.3835/plantgenome2011.08.0024}{Endelman 2011}.
}
\keyword{internal}
