% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_description.R
\docType{data}
\name{mphapdose}
\alias{mphapdose}
\title{Haplotype dosages}
\format{
A matrix (integer or character) indicating the haplotype present at each
individual homologue. Each row represents a locus and each column an individual
homologue.
}
\usage{
mphapdose
}
\description{
A matrix of simulated haplotype dosages for a tetraploid multiparental population
of 460 individuals. Since we simulated tetraploid individuals, the genotype of
each individual is represented by four columns, reporting the haplotype present at
each homologue.
}
\keyword{datasets}
