% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_description.R
\docType{data}
\name{mpmap}
\alias{mpmap}
\title{Genetic map}
\format{
A data frame with three columns for locus name, chromosome
and position.
}
\usage{
mpmap
}
\description{
A simulated genetic map of 1013 loci, indicating marker name, chromosome
and position. Markers in chromosome 0 have unknown position.
}
\keyword{datasets}
