% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_description.R
\docType{data}
\name{mpped}
\alias{mpped}
\title{Pedigree}
\format{
A character matrix with three columns (for offspring, parent 1
and parent 2) and 460 rows.
}
\usage{
mpped
}
\description{
A table of pedigrees (offspring, parent 1 and parent 2) for a simulated
tetraploid multiparental population.
}
\keyword{datasets}
