% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_description.R
\docType{data}
\name{mppheno}
\alias{mppheno}
\title{Phenotypes}
\format{
A vector of 460 elements.
}
\usage{
mppheno
}
\description{
A vector of simulated phenotypes for a tetraploid multiparental population.
}
\keyword{datasets}
