% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_description.R
\docType{data}
\name{mpsnpdose}
\alias{mpsnpdose}
\title{SNP dosages}
\format{
An integer matrix of SNP dosages, with markers in rows and
individuals in columns.
}
\usage{
mpsnpdose
}
\description{
A matrix of simulated SNP dosages for a tetraploid multiparental population
of 460 individuals.
}
\keyword{datasets}
