% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{original_order}
\alias{original_order}
\title{Original order}
\usage{
original_order(y, X)
}
\arguments{
\item{y}{phenotype matrix}

\item{X}{dosage matrix}
}
\value{
a vector that can be used to reorder the output of test.compatibility()
}
\description{
calc.Hinv calculates Hinv for every set of different dimensions
that are present in the data to improve speed (certain calculations
only depend on the dimensionality of the data). As a result, the Hinv
matrices are returned in an order that does not correspond the phenotype
order that is provided, causing incompatibilities later. This function
returns an index vector that can be used to reorder the Hinv matrix list.
}
\keyword{internal}
