% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{pairwise_rpool}
\alias{pairwise_rpool}
\title{Pairwise pooled correlation between multiallelic markers}
\usage{
pairwise_rpool(hapdos, ploidy)
}
\arguments{
\item{hapdos}{a list of matrices with allele dosages. Each matrix contains
allele dosages of one multiallelic marker, with alleles in rows and
individuals in columns.}

\item{ploidy}{numeric indicating the ploidy level.}
}
\value{
a correlation matrix between multiallelic markers
}
\description{
\code{pairwise_rpool} generates a pairwise correlation matrix
between a set of multiallelic markers. It is based on \code{rpool}.
}
\keyword{internal}
