% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{pcoa.plot}
\alias{pcoa.plot}
\title{PCoA plotter}
\usage{
pcoa.plot(
  K,
  comp = c(1, 2),
  plot_legend = T,
  col = NULL,
  coltype = NULL,
  h = NULL,
  l = NULL,
  legspace = 0.1,
  legname = NULL,
  pch = 19,
  ...
)
}
\arguments{
\item{K}{distance matrix}

\item{comp}{Vector of two integers, indicating which principal comonents to plot.
1 and 2 by default.}

\item{col}{vector for each individual in the matrix, indicating
a grouping. For instance, c("parent1","pop1",...). Each element will be
used to produce a colour.}

\item{coltype}{Argument to be passed to \code{\link{select.col}}.Either "sequential",
"qualitative", "divergent" or "rainbow".For a few categories, such as different treatments,
choose "qualitative" or "rainbow". For ordered categories, such as increasingly high levels
of a compound, use "sequential". For a gradient between two opposites, chose "divergent".}

\item{h}{Argument to be passed to \code{\link{select.col}}. It indicates the hue of
the colour palette to use.}

\item{l}{Value between 0 and 100 (default 60). Represents brightness, or the amount of white}

\item{legspace}{Numeric indicating the proportion of space to be left for plotting the legend
to the right of the plot. By default takes value 0.1 (10% of the x-value range). If legend
names are very long, increase this number to tweak the amount of space left to the right.}

\item{pch}{numeric vector containing 1 or as many values as unique values in "col"}

\item{...}{Additional parameters to be passed to the \code{plot} or \code{points}
method.}

\item{add}{Logical, whether to add the points to the current plot (using points()),
or make a new plot.}

\item{legend}{Logical, whether legend should be plotted.}
}
\value{
calculates and plots a pcoa plot
}
\description{
It produces a principal component plot for the K
distance matrix based on the prcomp() function.
}
\examples{

select.col(10,coltype = "qualitative")
select.col(10,coltype = "sequential", h = 280)
}
