% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{plot.LD}
\alias{plot.LD}
\title{Method for LD plotting}
\usage{
\method{plot}{LD}(LD, max_dist = NULL, main = NULL)
}
\arguments{
\item{LD}{LD list object as generated by \code{LD_decay}}

\item{max_dist}{numeric, maximum distance to plot}

\item{main}{character, title for the plot}
}
\value{
creates a plot
}
\description{
Given an LD list object (obtained via \code{LD_decay} function), it creates an LD
decay plot using all the percentiles calculated in \code{LD_decay}.
}
