% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{QQcalc}
\alias{QQcalc}
\title{QQ plot calculation}
\usage{
QQcalc(pvals)
}
\arguments{
\item{pvals}{a vector of p-values.}
}
\value{
observed and expected p-values
}
\description{
QQ plot calculation
}
\section{Functions}{
\itemize{
\item \code{QQcalc}: Used to calculate the expected p-values and the
sorted observed p-values. Automatically applies -log10 on the p-values.
}}

\keyword{internal}
