% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fun.R
\name{refineBlocks}
\alias{refineBlocks}
\title{Re-define haploblocks exceeding a maximum number of markers}
\usage{
refineBlocks(
  hb_list,
  nmrk = 2:8,
  mrkDosage = NULL,
  method = "split",
  nrand = 3
)
}
\arguments{
\item{hb_list}{haploblock list created by \code{map2blocks}}

\item{nmrk}{numeric vector indicating the allowed numbers of markers per
block}

\item{mrkDosage}{not used yet}

\item{method}{character indicating the method. Choose between c("split",
"random"):

* `split` will subset the block into blocks with maximum x SNP markers
(where x is max(nmrk)), according to the marker order provided.
* `random` will randomly select x SNP markers (where x is max(nmrk)).}

\item{nrand}{numeric. If \code{(method = "random")}, \code{nrand} indicates
the number of subsets (blocks) randomly extracted.}
}
\value{
a list where each element contains the names of markers grouped
in a block.
}
\description{
Re-define haploblocks exceeding a maximum number of markers
}
