% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{rpool}
\alias{rpool}
\title{Pooled correlation between two multiallelic markers}
\usage{
rpool(A, B, ploidy)
}
\arguments{
\item{A}{matrix of haplotype dosages of block A. Haplotypes in rows and
individuals in columns.}

\item{B}{matrix of haplotype dosages of block B.}

\item{ploidy}{numeric indicating the ploidy level.}
}
\value{
pooled correlation
}
\description{
\code{rpool} computes a pooled correlation coefficient between
two sets of variables (e.g. two multiallelic markers) using the equation in
Zhao et al. 2005, Genet. Res. 86:77-87.
Correlation of each pair of alleles is weighted by allele frequency.
}
\keyword{internal}
