% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{select.col}
\alias{select.col}
\title{Color selection based on HCL model}
\usage{
select.col(n, coltype = NULL, h = NULL, c = 100, l = NULL, alpha = 1)
}
\arguments{
\item{n}{Integer. Number of colours to return.}

\item{coltype}{Either "sequential", "qualitative" (default), "divergent" or "rainbow".
For a few categories, such as different treatments, choose "qualitative" or "rainbow".
For ordered categories, such as increasingly high levels of a compound, use
"sequential". For a gradient between two opposites, chose "divergent".}

\item{h}{One or two values between 0 and 360 (degrees in the colour wheel) to represent hue.
Default is c(120,240).#' If "qualitative" or "divergent" is used, the two hues will correspond
to each end of the colour
gradient. Otherwise only the first hue will be used. For reference, 0 is red, 140 is green, 240
is blue 300 is purple and 360 is back to red.}

\item{c}{Value between 0 and 100 (default 100). Represents chroma, or colour intensity and is similar to saturation.
0 means grey, 100 means completely saturated.}

\item{l}{Value between 0 and 100 (default 60). Represents brightness, or the amount of white}

\item{alpha}{Value between 0 (transparent) and 1 (opaque), to add some degree of
transparency to the colour palette.}
}
\value{
A colour palette of n colours.
}
\description{
Choice of color palettes can have a relevant impact not only
in the outlook of data visualizations, but also on their interpretation.
There exist many models of colour description, such as rgb, or cymk, but in
this package we chose the hcl model (Hue, Chroma, Luminance). Hue is expressed as
an angle (usually between 0 and 360) in the colour wheel; Chroma, between0 and 100,
reflects the intensity of the pigment (similar to saturation); and
Lightness, the amount of white/black in a colour. Such organization of colour
aligns very well with human colour perception and thus is more intuitive to
use than rgb or cymk models, where the colour outcome is not so intuitive.
For this reason, the package \code{colorspace} has been developed, which helps
in generating pleasant and adequate palettes for data visualization. For more
information on how to choose colour palettes be sure to visit the
\href{http://colorspace.r-forge.r-project.org/articles/hcl_palettes.html}{colorspace blog}. Or
if you whish to learn about colour spaces and the HCL colour space visit
\href{http://hclwizard.org/why-hcl/}{"why HCL"}.
}
\examples{

select.col(10,coltype = "qualitative")
select.col(10,coltype = "sequential", h = 280)
}
