% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{stndrdz.pheno}
\alias{stndrdz.pheno}
\title{Standaridze phenotypes}
\usage{
stndrdz.pheno(phenotypes)
}
\arguments{
\item{phenotypes}{numeric matrix}
}
\description{
Given a matrix of numeric values, returns a list with
the mean and sd per column as well as standaridzed matrix
}
\keyword{internal}
