% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{test.compatibility}
\alias{test.compatibility}
\title{Comparison of dimensions between matrices of a mixed model.}
\usage{
test.compatibility(y, X, Z, K, Hinv = NULL)
}
\arguments{
\item{y}{matrix of response y}

\item{X}{matrix of fixed effects X}

\item{Z}{matrix of random effects Z}

\item{K}{matrix of variance structure K}
}
\value{
a list of dimension-compatible matrices without missing values.
}
\description{
Dimensions of \eqn{y}, \eqn{X}, \eqn{K} and \eqn{Z}
matrices are checked to ensure that
they are mathematically compatible. That is to say: \eqn{y} is a vector of
length \eqn{n}, \eqn{X} is a matrix of \eqn{n*m} where \eqn{m} is the number of parameters,
\eqn{K} is a square matrix of \eqn{n*n} and Z is a matrix of
\eqn{k*n}.

This internal function is not intended for general use.
}
\keyword{internal}
