% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{thr.LiJi}
\alias{thr.LiJi}
\title{Calculate significance level based on the effective number of independent tests}
\usage{
thr.LiJi(m, chrom, alpha = 0.05, ploidy = NULL)
}
\arguments{
\item{m}{A matrix of SNP genotypes (discrete or continuous) with markers
in rows and individuals in columns.}

\item{chrom}{A vector defining to which chromosome markers belong.}

\item{alpha}{Significance level (use 0.05 for 5\% significance).}

\item{ploidy}{Numeric indicating the ploidy level.}
}
\value{
Value of adjusted alpha.
}
\description{
This function implements a method to adjust the significance
level for multiple testing based on the effective number of independent
tests, as proposed by \href{https://www.nature.com/articles/6800717?report=reader}{Li and Ji (2005, Heredity)}.
The Bonferroni adjusted threshold is obtained dividing a significance level
by the number of tests, assuming tests are independent to each other.
The Li and Ji method is more appropriate when the assumption of
independence is not met. In this method, the eigenvalues of a correlation
matrix are used to estimate the effective number of independent tests.
}
\examples{
## Get example SNP dosages and a map
data("mpsnpdose")
data("mpmap")

mpsnpdose[1:5,1:5]
head(mpmap)
tapply(mpmap$marker, mpmap$chromosome, length)
identical(rownames(mpsnpdose), mpmap$marker)

## excluding unmapped markers (chromosome 0)
mapped <- mpmap$chromosome != 0
thr <- thr.LiJi(m = mpsnpdose[mapped,],
                chrom = mpmap$chromosome[mapped])
-log10(thr$threshold) #log scale
}
