% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{tick_wheel}
\alias{tick_wheel}
\title{Draws a tick wheel}
\usage{
tick_wheel(n, cen = c(0, 0), r = 1)
}
\arguments{
\item{n}{number of ticks}

\item{cen}{numeric, x and y coordinates to locate the center of the circle}

\item{r}{numeric, radius of the circle}
}
\value{
plots ticks around a center
}
\description{
Draws a tick wheel
}
\keyword{internal}
