% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{calc.K}
\alias{calc.K}
\title{Calculation of realized distance matrix (K)}
\usage{
calc.K(matrix, haplotypes = FALSE, ploidy = NULL)
}
\arguments{
\item{matrix}{A matrix of SNP genotypes (continuous or discrete) or
haplotypes (multi-allelic markers), where columns are markers and
rows may be either (1) individuals (for biallelic markers) or (2)
individual homologues (for multi-allelic markers, such as haplotypes).}

\item{haplotypes}{Logical, whether haplotypes are present in the matrix. If TRUE,
a matrix is expected to have p rows per individual, where p is ploidy.}

\item{ploidy}{integer, the ploidy number p of all individuals. Only used if
haplotypes = TRUE.}
}
\value{
A numeric matrix nxn where n is the number of rows.
}
\description{
Using SNP genotypes (discrete or continuous) or haplotype
scores, a distance matrix is calculated
such that the average distance of an individual with itself is 1, and
the average with an unrelated individual is 0. Based on the "Realized
Relationship" model found in \href{https://dl.sciencesocieties.org/publications/tpg/abstracts/9/2/plantgenome2015.08.0073}{Rosyara et al. 2016}
}
\examples{
## Create SNP dosages for 10 tetraploid individuals (in rows) and 100 markers
snpdose <- matrix(sample(0:4,1000, replace=TRUE), nrow=10)

## Calculate K matrix
K <- calc.K(snpdose)

## Create an haplotype matrix for 5 tetraploid individuals (in rows) and 100 markers
## Each individual genotype is represented by four rows (its four homologues)
hapdose <- matrix(sample(1:6,2000, replace=TRUE), nrow=20)

## K matrix
K <- calc.K(hapdose, haplotype = TRUE, ploidy = 4)

## Haplotypes can be named by characters as well
hapdose <- matrix(letters[hapdose], nrow=20)
K <- calc.K(hapdose, haplotype = TRUE, ploidy = 4)

}
