% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{dosage.X}
\alias{dosage.X}
\title{X matrix calculator}
\usage{
dosage.X(genotypes, haplotype = FALSE, ploidy = NULL, normalize = FALSE)
}
\arguments{
\item{genotypes}{vector of dosages per individual, or p haplotypes per individual
(where p is ploidy).}

\item{haplotype}{logical, whether the vector contains haplotype or genotype data}

\item{ploidy}{integer indicating ploidy.}

\item{normalize}{logical, should the X matrix be normalized?}
}
\value{
if haplotypes = FALSE, a matrix is returned whith a single column
containing the same values that were provided. If haplotypes = TRUE, a design matrix
is created with ncol = number of haplotypes and nrow = number of individuals.
}
\description{
Using a vector of dosages per individual, or p haplotypes per
individual (where p is ploidy), it creates a design matrix X, with or without
normalization.
}
\examples{

## SNP dosages of 10 tetraploid individuals
snpdose <- sample(0:4, 10, replace = TRUE)
dosage.X(snpdose)

## Haplotypes of 10 tetraploid individuals for a locus with 5 haplotypes
hapdose <- sample(1:5,40, replace = TRUE)
dosage.X(genotypes = hapdose, haplotype = TRUE, ploidy = 4)
}
\keyword{internal}
