% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{draw_chrom_axis}
\alias{draw_chrom_axis}
\title{Chromosome axis drawer}
\usage{
draw_chrom_axis(ch_start, ch_end, chrom, small = F)
}
\arguments{
\item{ch_start}{vector of chromosome starts}

\item{ch_end}{vector of chromosome ends}

\item{chrom}{names of chromosomes}

\item{small}{logical, should small axis be also drawn?}
}
\value{

}
\description{

}
\keyword{internal}
