% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fun.R
\name{HapdoseToHapname}
\alias{HapdoseToHapname}
\title{Convert haplotype dosages into haplotype names}
\usage{
HapdoseToHapname(hapdose, ploidy)
}
\arguments{
\item{hapdose}{Either the output of PolyHaplotyper or a list where each element
is a numeric matrix of haplotype dosages for a certain locus and each element name
is a locus name. Locus names must be unique.
In each matrix, haplotypes are in rows and individuals in columns.
Row names must contain haplotype
names.
Column names contain individual names.}

\item{ploidy}{numeric indicating the ploidy level (identical for all the
individuals).}
}
\value{
a matrix of multiallelic genotypes, with loci (blocks) in rows and
individual homologues in columns. Genotypes are indicated by using haplotype
names.
}
\description{
A genotype of an individual can be represented either by a
numerical vector of dosages of each haplotype or by a numerical vector of
haplotype names (numbers) for each homologue. In the former case, vector
length will correspond to the number of haplotypes, while in the latter
case it will be the number of homologues, that is the individual ploidy.
This function converts a list of haplotype dosages into a matrix of
haplotypes names.
}
