% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{impute.knn}
\alias{impute.knn}
\title{KNN genotype imputator}
\usage{
impute.knn(geno, ploidy, map = NULL, kneighbors = 30, K = NULL)
}
\arguments{
\item{geno}{Haplotype (accepts characters) or Dosage matrix}

\item{ploidy}{Ploidy of the organism}

\item{map}{Optional, genetic map used in \code{\link{sample.cM}}.
Improves efficiency}

\item{kneighbors}{Number of nearest neighbours to use in the imputation. Default is 20.}

\item{K}{Optionally, provide the similarity matrix (K) directly.}
}
\value{
A Haplotype or Dosage matrix without missing values.
}
\description{
K-Nearest Neighbour imputation algorithms are useful for
imputing missing data. In this function, we use the kinship matrix
calculated with \code{\link{calc.K}} to determine the most similar
individuals, and substitute the missing values of one individual with
the most common dosage/haplotype of its nearest neighbours.
}
