% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{inputCheck_K}
\alias{inputCheck_K}
\title{Check input requirements for the argument K of map.QTL}
\usage{
inputCheck_K(K)
}
\arguments{
\item{K}{the input of the argument K of map.QTL}
}
\value{
When requirements are met, it returns a list with:
* $type: number specifying the input type found: 1 for NULL, 2 for TRUE
 or 3 for a numeric square matrix;
* $K: the K object.
}
\description{
Check that the argument K is either 1) NULL, 2) TRUE or
3) a numeric square matrix.
}
\keyword{internal}
