% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{inputCheck_Q}
\alias{inputCheck_Q}
\title{Check input requirements for the argument Q of map.QTL}
\usage{
inputCheck_Q(Q)
}
\arguments{
\item{Q}{the input of the argument Q of map.QTL}
}
\value{
When requirements are met, it returns a list with:
* $type: number specifying the input type found: 1 for NULL, 2 for TRUE,
 3 for a population membership vector or 4 for a design matrix;
* $Q: the Q object.
}
\description{
Check that the argument Q is either 1) NULL, 2) TRUE,
3) a vector (num or char) indicating sub-population membership or 4) a user
provided design matrix.
}
\keyword{internal}
