% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{map.QTL}
\alias{map.QTL}
\title{QTL mapping of a matrix of phenotypes}
\usage{
map.QTL(
  phenotypes,
  genotypes,
  ploidy,
  map,
  K = NULL,
  Q = NULL,
  Z = NULL,
  cofactor = NULL,
  cofactor.type = NULL,
  cM = 1,
  seed = NULL,
  Qpco = 2,
  no_cores = parallel::detectCores() - 1,
  approximate = T,
  permutation = NULL,
  fam = NULL,
  nperm = 100,
  alpha = 0.05,
  impute = T,
  k = 20,
  linear = NULL,
  K_identity = F
)
}
\arguments{
\item{phenotypes}{A numeric matrix of phenotypes,
rows are individuals and columns are different phenotypes.
Not updated for vector phenotypes yet. Must follow same
order of individuals as genotypes.}

\item{genotypes}{A matrix of genotypes, rows are markers and
columns may be either (1) individual dosages of biallelic markers, with column
names coinciding with phenotype individual names or (2) chromosome alleles of each
individual (#gt: for multiallelic markers, such as haplotypes).
Must follow same order of individuals as phenotypes.}

\item{ploidy}{a number indicating the ploidy level. All the individuals
must have same ploidy.}

\item{map}{A data frame with 3 columns containing map information. The first
column specifies marker names, the second column chromosome names and the
third one marker position (any map unit).
Marker position is used for plotting or it could be used to sample a subset
of evenly spaced markers to be used for kinship estimation.}

\item{K}{NULL, T or distance matrix. If NULL, no relatedness
matrix will be used (i.e., a linear model will be applied). If T
a K distance matrix will be calculated. A distance matrix may also
be directly specified.}

\item{Q}{NULL, T or vector identifying populations or Q design matrix (#gt: explain it). If NULL, no Q will
be included in the model (i.e., a model without Q correction). If T, a pco
decomposition will be used to estimate population differentiation. If a
vector specifying population of each individual is passed,
it will be used to construct a Q matrix. Vector may contain numerical
or character.}

\item{Z}{Identity matrix indicating which individuals correspond to which
genotypic effects. For instance, if multiple samples correspond to the same
individual, this matrix should indicate so. #gt: shouldn't it be internal only?}

\item{cofactor}{Possible cofactor matrix (where each column is a cofactor).}

\item{cofactor.type}{If a cofactor matrix is specified, a character vector
specifying the type ("numerical" or "categorical") of each cofactor.
It accepts partial strings such as "cat" and "num".}

\item{cM}{Numeric. K distance matrix (#gt: and Q matrix) will be calculated using markers every
cM centiMorgans (#gt: we should change the parameter name, since the map can be in any unit). Defaults to 1.}

\item{seed}{An integer to set a seed for random number generation in \code{sample.cM}.}

\item{Qpco}{Logical value indicating whether a PCo-based population factor should be
estimated and included. Is useful for detecting and correcting for population
structure.}

\item{no_cores}{Numeric. Number of cores to be used in parallel computing
of the p-values. Defaults to number of cores -1.}

\item{approximate}{Logical value indicating whether P3D/EMMAX approach should be
used for computational efficiency (if F, expect much longer waiting times)}

\item{permutation}{permutation strategy. Can be "pop" (permutation over the
whole population) or "fam" (permutation within families). If it is NULL, no
permutation will be run.}

\item{fam}{If permutation = "fam", provide here a vector indicating the family
membership (or a population structure membership) of each individual.}

\item{nperm}{number of permutations, if \code{permutation} is not null.}

\item{alpha}{Permutation threshold alpha value, default to 0.05. Alpha is the
chance of observing a false positive experiment-wise.}

\item{impute}{Logical value indicating whether missing genotypes should be
imputed using \code{impute.knn}. Defaults to True, but False is recommended (imputation algorithm
needs to be improved and with few missing values it has a small effect
on QTL detection)}

\item{k}{Number of neighbours to use in the internal function \code{impute.knn}}

\item{linear}{logical. If True, linear model (without structure correction)
is applied. If False, mixed model (with structure correction) is applied. If
not specified, it looks at the value of K, and only applied linear model
if K = NULL.}

\item{K_identity}{logical. If TRUE, an identity matrix is used in the
random term of a mixed model. This is to run a mixed model with no kinship
correction, even if a kinship matrix is provided or calculated for other
purposes (e.g. imputation of missing genotypes, calculation of the Q matrix)}
}
\value{
a pvalue matrix containing the pvalue of each marker with each phenotype passed.
}
\description{
Main function and wrapper for QTL analysis in polyploids using
biallelic or multiallelic markers. For a full description of the function please
see the official vignette.
}
