% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{map_axis}
\alias{map_axis}
\title{Axis calculator}
\usage{
map_axis(maplist, space = 0)
}
\arguments{
\item{maplist}{map dataframe (or list of map dataframes)}

\item{space}{space to leave between chromosomes (in map$position units)}
}
\value{
a maplist with the original maps and an extra "axis" column, and
a data.frame with the axis chromosomes start and end, for axis plotting.
}
\description{
Using a map dataframe, it calculates the
position of each marker onto a plot, including
spaces between chromosomes, if desired.
}
\keyword{internal}
