% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{plot.LD}
\alias{plot.LD}
\title{Plotting linkage disequilibrium (LD) decay}
\usage{
\method{plot}{LD}(LD, max_dist = NULL, main = NULL)
}
\arguments{
\item{LD}{LD list object as generated by \code{LD_decay}.}

\item{max_dist}{numeric, maximum distance to plot.}

\item{main}{character, title for the plot.}
}
\value{
creates a plot.
}
\description{
Given an LD list object (obtained via \code{LD_decay} function), it creates an LD
decay plot using all the percentiles calculated in \code{LD_decay}.
}
\examples{
# ## Get example SNP dosages and a map
data("mpsnpdose")
data("mpmap")

mpsnpdose[1:5,1:5]
head(mpmap)
tapply(mpmap$marker, mpmap$chromosome, length)
identical(rownames(mpsnpdose), mpmap$marker)

## LD decay calculation
LD <- LD_decay(mpsnpdose, mpmap, win_size = 1,
               max_dist = 100, percentile = c(0.9,0.95))

## plot LD decay
\dontrun{plot_LD(LD, max_dist = NULL, main=NULL)}

}
