% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{sample.cM}
\alias{sample.cM}
\title{Sample markers every X centimorgans}
\usage{
sample.cM(genotypes, map, cM = 1, seed = NULL)
}
\arguments{
\item{genotypes}{Either a vector where each element is a marker, or a matrix
where each row is a marker.}

\item{map}{table containing a "chromosome" (numerical) and a "position" (in cM) column.
Markers with chromosome number == 0 will not be taken into account}

\item{cM}{Size of bin to sample a marker from.}

\item{seed}{An integer to set a seed for random number generation.}
}
\value{
A table as genotypes, with as many rows as can be obtained by sampling
markers every X centimorgans.
}
\description{
Obtains a subset of markers homogeneously distributed across
a genetic map based on a cM distance. For instance, if cM=1, will get 1 marker
for every 1 cM window if it can. Markers added on chromosome 0 are not returned
as their genetic position is not known (and thus cannot be cM sampled)
}
