% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fun.R
\name{skyplot}
\alias{skyplot}
\title{Manhattan Skyline Plot}
\usage{
skyplot(
  pval,
  map,
  threshold = NULL,
  ylab = NULL,
  xlab = NULL,
  ylim = NULL,
  chrom = NULL,
  small = NULL,
  col = NULL,
  h = NULL,
  l = NULL,
  pch = NULL,
  ...
)
}
\arguments{
\item{pval}{Numerical vector. Usually -log10(p-values), but other values are accepted.}

\item{map}{Dataframe containing columns "chromosome" and "position"}

\item{threshold}{A threshold value to draw the threshold line.}

\item{ylab}{character, label to add on the y axis, defaults to "-log10(pval)"}

\item{xlab}{character, label to add on the x axis, defaults to "Chromosome"}

\item{ylim}{numeric vector of length two, minimum and maximum y axis values}

\item{chrom}{A vector of chromosome names to be included in the plot}

\item{small}{logical, should the small axis (per chromosome) be drawn? By default
T only if number of chromosomes <3.}

\item{col}{Base colour to use for plotting. Odd chromosomes will be plotted with this colour,
even chromosomes will be plotted with a lighter version of the same colour.}

\item{h}{One or two values between 0 and 360 (degrees in the colour wheel) to represent hue.
Default is c(120,240).#' If "qualitative" or "divergent" is used, the two hues will correspond
to each end of the colour
gradient. Otherwise only the first hue will be used. For reference, 0 is red, 140 is green, 240
is blue 300 is purple and 360 is back to red.}

\item{l}{Value between 0 and 100 (default 60). Represents brightness, or the amount of white}

\item{pch}{numeric indicating the type of point to be passed to plot()}

\item{...}{Other parameters to be passed to plot()}
}
\value{
creates a skyline plot
}
\description{
Takes a vector of values and a dataframe of a genetic map containing the columns
"chromosome" and "position", and produces a plot where values are mapped onto the genome.
}
