% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQTL_fun.R
\name{thr.LiJi}
\alias{thr.LiJi}
\title{Li and Ji method for significance level in multiple testing}
\usage{
thr.LiJi(m, chrom, alpha = 0.05, ploidy)
}
\arguments{
\item{m}{a matrix of SNP dosages or a list of matrices with haplotype dosages,
with markers (or haplotypes) in rows and individuals in columns.}

\item{chrom}{a vector defining to which chromosome markers belong.}

\item{alpha}{significance level (use 0.05 for 5\% significance).}

\item{ploidy}{numeric indicating the ploidy level.}
}
\value{
value of adjusted alpha.
}
\description{
This function implements a method to adjust the significance
level for multiple testing, as proposed by Li and Ji (Li and Ji, 2005, Heredity).
The Bonferroni adjusted threshold is obtained dividing a significance level
by the number of tests, assuming tests are independent to each other.
The Li and Ji method is more appropriate when the assumption of
independence is not met. In this method, the eigenvalues of a correlation
matrix are used to estimate the effective number of independent tests.
NOT RECOMMENDED fo haplotype data, use SNP dosages only.
}
